+++ src/components/Controls.vue
@@ -238,6 +238,9 @@ export default {
 			}
 		},
 		async onBulkDelete() {
+			if (!confirm(t('bookmarks', 'Do you really want to delete these items?'))) {
+				return
+			}
 			await this.$store.dispatch(actions.DELETE_SELECTION, { folder: this.$route.params.folder })
 			this.$store.commit(mutations.RESET_SELECTION)
 		},
